+++ MR-Solution/MedicalResearch.Domain/Services/VisitService.cs
@@ -1,5 +1,6 @@
 ﻿using MedicalResearch.DAL.UnitOfWork;
 using MedicalResearch.Domain.Exceptions;
+using MedicalResearch.Domain.Extensions;
 using MedicalResearch.Domain.Interfaces.Service;
 using MedicalResearch.Domain.Models;
 using MedicalResearch.Domain.Queries;
@@ -11,60 +12,61 @@ public class VisitService(IUnitOfWork unitOfWork, ILogger<VisitService> logger) 
 {
     public async Task<Visit> AddVisitAsync(Visit visit)
     {
+        Visit? added;
+        int countAdded;
+        ClinicStockMedicine? clinicStockMedicineUpdated;
+
+        var medicine = await unitOfWork.ClinicStockMedicineRepository.GetClinicStockMedicineAsync(visit.MedicineId, visit.ClinicId) ?? throw new DomainException("Medicine not found");
+        var clinic = await unitOfWork.ClinicRepository.GetByIdAsync(visit.ClinicId) ?? throw new DomainException("Clinic not found");
+        var existedVisits = await unitOfWork.VisitRepository.GetVisitsOfPatient(visit.PatientId);
+        var existedVisit = existedVisits.FirstOrDefault(x => x.ClinicId.Equals(visit.ClinicId) && x.DateOfVisit.Equals(visit.DateOfVisit));
+        if (existedVisit != null)
+        {
+            throw new DomainException("Visit already exists");
+        }
+        if (medicine.Amount <= 0)
+        {
+            throw new DomainException("Not enough medicine in stock");
+        }
         try
         {
-            var medicine = await unitOfWork.ClinicStockMedicineRepository.GetClinicStockMedicineAsync(visit.MedicineId, visit.ClinicId) ?? throw new DomainException("Medicine not found");
-            var clinic = await unitOfWork.ClinicRepository.GetByIdAsync(visit.ClinicId) ?? throw new DomainException("Clinic not found");
-            var existedVisits = await unitOfWork.VisitRepository.GetVisitsOfPatient(visit.PatientId);
-            var existedVisit = existedVisits.FirstOrDefault(x => x.ClinicId.Equals(visit.ClinicId) && x.DateOfVisit.Equals(visit.DateOfVisit));
-            if (existedVisit != null)
-            {
-                throw new DomainException("Visit already exists");
-            }
-            if (medicine.Amount <= 0)
-            {
-                throw new DomainException("Not enough medicine in stock");
-            }
-            else
-            {
-                medicine.Amount -= 1;
-                unitOfWork.ClinicStockMedicineRepository.Update(medicine);
-            }
+            medicine.Amount -= 1;
+            clinicStockMedicineUpdated = unitOfWork.ClinicStockMedicineRepository.Update(medicine);
+
             var numberOfVisit = unitOfWork.VisitRepository.GetNumberOfNextVisit(visit.PatientId);
             visit.NumberOfVisit = numberOfVisit;
-            var added = await unitOfWork.VisitRepository.AddAsync(visit);
-            return await unitOfWork.SaveAsync() > 0 ? added : throw new DomainException("Visit not added");
-        }
-        catch (DomainException ex)
-        {
-            logger.LogError(ex, "Visit {visit} could not be added: {message}", visit.Id, ex.Message);
-            throw;
+            added = await unitOfWork.VisitRepository.AddAsync(visit);
+            countAdded = await unitOfWork.SaveAsync();
         }
         catch (Exception ex)
         {
             logger.LogError(ex, "Visit {visit} could not be added: {message}", visit.Id, ex.Message);
             throw new DomainException($"Error while adding Visit {visit.Id}");
         }
+        return countAdded > 0 && added != null && clinicStockMedicineUpdated != null ? added : throw new DomainException("Visit not added");
     }
 
     public async Task<bool> DeleteVisitAsync(int id)
     {
+        ClinicStockMedicine? clinicStockMedicineUpdated;
+        int countDeleted = 0;
+        bool isDelete = false;
+
+        var visit = await unitOfWork.VisitRepository.GetByIdAsync(id) ?? throw new DomainException("Visit not found");
+        var medicine = await unitOfWork.ClinicStockMedicineRepository.GetClinicStockMedicineAsync(visit.MedicineId, visit.ClinicId) ?? throw new DomainException("Medicine not found");
         try
         {
-            var visit = await unitOfWork.VisitRepository.GetByIdAsync(id) ?? throw new DomainException("Visit not found");
-            var medicine = await unitOfWork.ClinicStockMedicineRepository.GetClinicStockMedicineAsync(visit.MedicineId, visit.ClinicId) ?? throw new DomainException("Medicine not found");
-            if (medicine != null)
+            medicine.Amount += 1;
+            clinicStockMedicineUpdated = unitOfWork.ClinicStockMedicineRepository.Update(medicine);
+            if (clinicStockMedicineUpdated != null)
             {
-                medicine.Amount += 1;
-                unitOfWork.ClinicStockMedicineRepository.Update(medicine);
+                isDelete = unitOfWork.VisitRepository.Delete(visit);
             }
-            var isDelete = unitOfWork.VisitRepository.Delete(visit);
-            return await unitOfWork.SaveAsync() > 0 ? isDelete : throw new DomainException("Visit not deleted");
-        }
-        catch (DomainException ex)
-        {
-            logger.LogError(ex, "Visit with id {id} could not be deleted: {message}", id, ex.Message);
-            throw;
+            if (isDelete)
+            {
+                countDeleted = await unitOfWork.SaveAsync();
+            }
+            return countDeleted > 0 && isDelete == true &&  clinicStockMedicineUpdated != null ? isDelete : throw new DomainException("Visit not deleted");
         }
         catch (Exception ex)
         {
@@ -86,7 +88,7 @@ public class VisitService(IUnitOfWork unitOfWork, ILogger<VisitService> logger) 
         }
     }
 
-    public async Task<List<Visit>> GetVisitsAsync(Query query)
+    public async Task<PagedList<Visit>> GetVisitsAsync(Query query)
     {
         try
         {
@@ -101,28 +103,30 @@ public class VisitService(IUnitOfWork unitOfWork, ILogger<VisitService> logger) 
 
     public async Task<Visit> UpdateVisitAsync(Visit visit)
     {
+        Visit? updated;
+        int countUpdated = 0;
+
+        var existingVisit = await unitOfWork.VisitRepository.GetByIdAsync(visit.Id) ?? throw new DomainException("Visit not found");
+        var medicine = await unitOfWork.ClinicStockMedicineRepository.GetClinicStockMedicineAsync(visit.MedicineId, visit.ClinicId) ?? throw new DomainException("Medicine not found");
+        var existedVisits = await unitOfWork.VisitRepository.GetVisitsOfPatient(visit.PatientId);
+        var existedNumber = existedVisits.FirstOrDefault(x => x.ClinicId.Equals(visit.ClinicId) && x.PatientId.Equals(visit.PatientId) && x.NumberOfVisit.Equals(visit.NumberOfVisit));
+        if (existedNumber != null)
+        {
+            throw new DomainException("Number of visit already exists");
+        }
         try
         {
-            var existingVisit = await unitOfWork.VisitRepository.GetByIdAsync(visit.Id) ?? throw new DomainException("Visit not found");
-            var medicine = await unitOfWork.ClinicStockMedicineRepository.GetClinicStockMedicineAsync(visit.MedicineId, visit.ClinicId) ?? throw new DomainException("Medicine not found");
             existingVisit.DateOfVisit = visit.DateOfVisit;
             existingVisit.ClinicId = visit.ClinicId;
             existingVisit.PatientId = visit.PatientId;
             existingVisit.MedicineId = visit.MedicineId;
-            var existedVisits = await unitOfWork.VisitRepository.GetVisitsOfPatient(visit.PatientId);
-            var existedNumber = existedVisits.FirstOrDefault(x => x.ClinicId.Equals(visit.ClinicId) && x.PatientId.Equals(visit.PatientId) && x.NumberOfVisit.Equals(visit.NumberOfVisit));
-            if (existedNumber != null)
-            {
-                throw new DomainException("Number of visit already exists");
-            }
             existingVisit.NumberOfVisit = visit.NumberOfVisit;
-            var updated = unitOfWork.VisitRepository.Update(existingVisit);
-            return await unitOfWork.SaveAsync() > 0 ? updated : throw new DomainException("Visit not updated");
-        }
-        catch (DomainException ex)
-        {
-            logger.LogError(ex, "Visit {visit} could not be updated: {message}", visit.Id, ex.Message);
-            throw;
+            updated = unitOfWork.VisitRepository.Update(existingVisit);
+            if (updated != null)
+            {
+                countUpdated = await unitOfWork.SaveAsync();
+            }
+            return countUpdated > 0 && updated != null ? updated : throw new DomainException("Visit not updated");
         }
         catch (Exception ex)
         {
